move icurrentuser to common/interfaces
