+++ src/Core/Application/Startup.cs
@@ -7,8 +7,11 @@ namespace DN.WebApi.Application;
 
 public static class Startup
 {
-    public static IServiceCollection AddApplication(this IServiceCollection services) =>
-        services
-            .AddValidatorsFromAssembly(Assembly.GetExecutingAssembly())
-            .AddMediatR(Assembly.GetExecutingAssembly());
+    public static IServiceCollection AddApplication(this IServiceCollection services)
+    {
+        var assembly = Assembly.GetExecutingAssembly();
+        return services
+            .AddValidatorsFromAssembly(assembly)
+            .AddMediatR(assembly);
+    }
 }
