+++ src/Host/Controllers/Identity/IdentityController.cs
@@ -1,3 +1,4 @@
+using DN.WebApi.Application.Common.Interfaces;
 using DN.WebApi.Application.Identity.Users;
 using DN.WebApi.Application.Identity.Users.Password;
 using DN.WebApi.Infrastructure.Auth.Permissions;
