+++ src/Infrastructure/Catalog/BrandGeneratorJob.cs
@@ -2,7 +2,6 @@
 using DN.WebApi.Application.Catalog.Brands;
 using DN.WebApi.Application.Common.Interfaces;
 using DN.WebApi.Application.Common.Persistence;
-using DN.WebApi.Application.Identity.Users;
 using DN.WebApi.Domain.Catalog.Brands;
 using DN.WebApi.Shared.Notifications;
 using Hangfire;
