+++ src/Infrastructure/Middleware/ExceptionMiddleware.cs
@@ -1,7 +1,6 @@
 using System.Net;
 using DN.WebApi.Application.Common.Exceptions;
 using DN.WebApi.Application.Common.Interfaces;
-using DN.WebApi.Application.Identity.Users;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Localization;
 using Serilog;
