+++ src/Infrastructure/Persistence/Context/ApplicationDbContext.cs
@@ -1,7 +1,6 @@
 using System.Data;
 using DN.WebApi.Application.Common.Events;
 using DN.WebApi.Application.Common.Interfaces;
-using DN.WebApi.Application.Identity.Users;
 using DN.WebApi.Domain.Catalog.Brands;
 using DN.WebApi.Domain.Catalog.Products;
 using DN.WebApi.Domain.Common.Contracts;
